Reproducible KeyStore
Previously, when a KeyStore was stored (written out) the order that the
entries were written was non-deterministic, depending on the iteration order
of map, which in Go is both undefined and in practice, variable.  The result
of the non-deterministic ordering was that while multiple writes of the same
set of entries into a KeyStore would result in functionally equivalent files
on the file system, they actual bytes would not be identical.  For most
scenarios this doesn't matter, but for Docker image layers, it absolutely
does.  Without identical bytes, layer de-duplication and transfer avoidance
cannot be performed.

This change ensures that entries are always written into the KeyStore in a
deterministic order (lexically sorted by alias), and the result is
byte-identical to all other writes with the same entries.

Signed-off-by: Ben Hale <bhale@vmware.com>